utils/requirements.txt
python-whois
dnspython
requests
sslyze
python3-nmap
checkdmarc
snallygaster
rich
